requirements.txt
alpaca-py>=0.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
pytest>=8.0.0
